+++ src/main/java/insane96mcp/shieldsplus/module/base/BaseFeature.java
@@ -5,28 +5,22 @@ import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
-import insane96mcp.insanelib.data.IdTagValue;
 import insane96mcp.shieldsplus.ShieldsPlus;
 import insane96mcp.shieldsplus.setup.SPEnchantments;
-import insane96mcp.shieldsplus.setup.SPShieldMaterials;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
 import insane96mcp.shieldsplus.world.item.enchantment.*;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.item.ShieldItem;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.living.LivingHurtEvent;
 import net.minecraftforge.event.entity.living.ShieldBlockEvent;
-import net.minecraftforge.event.entity.player.ItemTooltipEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-
-import java.util.List;
 
 @Label(name = "Shields+")
 @LoadFeature(module = ShieldsPlus.RESOURCE_PREFIX + "base", canBeDisabled = false)
 public class BaseFeature extends Feature {
     @Config(min = 0)
-    @Label(name = "Shield Windup", description = "In vanilla when you start blocking with a shield, there's a 0.25 seconds (5 ticks) window where you are still not blocking. By default is now set to 0.1 seconds.")
-    public static Integer shieldWindup = 2;
+    @Label(name = "Shield Windup", description = "In vanilla when you start blocking with a shield, there's a 0.25 seconds (5 ticks) window where you are still not blocking. By default the windup is disabled.")
+    public static Integer shieldWindup = 0;
     @Config
     @Label(name = "Shields Block Fixed Damage Amount", description = "If true shields will block only a certain amount of damage. If false the vanilla behaviour is used.")
     public static Boolean shieldBlockFixedDamageAmount = true;
@@ -36,19 +30,9 @@ public class BaseFeature extends Feature {
     @Config
     @Label(name = "Combat Test shield disabling", description = "Makes shields always disable for 1.6 seconds like Combat Test snapshots.")
     public static Boolean combatTestShieldDisabling = true;
-
-    public static final List<IdTagValue> DEFAULT_BLOCKING_VALUES = List.of(
-            IdTagValue.newId("shieldsplus:wooden_shield", 1f),
-            IdTagValue.newId("shieldsplus:stone_shield", 1f),
-            IdTagValue.newId("shieldsplus:wooden_shield", 1f),
-            IdTagValue.newId("shieldsplus:wooden_shield", 1f),
-            IdTagValue.newId("shieldsplus:wooden_shield", 1f),
-            IdTagValue.newId("shieldsplus:wooden_shield", 1f),
-            IdTagValue.newId("shieldsplus:wooden_shield", 1f),
-            IdTagValue.newId("shieldsplus:wooden_shield", 1f),
-            IdTagValue.newId("shieldsplus:wooden_shield", 1f),
-            IdTagValue.newId("shieldsplus:wooden_shield", 1f)
-    );
+    @Config
+    @Label(name = "Lifted and Cooldown", description = "If true, shields can be lifted only for a certain amount of time and will go on cooldown.")
+    public static Boolean liftedAndCooldown = true;
 
     public BaseFeature(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);
@@ -56,16 +40,8 @@ public class BaseFeature extends Feature {
 
     @SubscribeEvent
     public void onShieldBlock(ShieldBlockEvent event) {
-        if (shieldBlockFixedDamageAmount) {
-            double baseBlockedDamage;
-            if (event.getEntity().getUseItem().is(Items.SHIELD)) {
-                baseBlockedDamage = SPShieldMaterials.IRON.damageBlocked;
-            }
-            else if (event.getEntity().getUseItem().getItem() instanceof SPShieldItem) {
-                baseBlockedDamage = ((SPShieldItem) event.getEntity().getUseItem().getItem()).getBlockedDamage(event.getEntity().getUseItem(), event.getEntity(), event.getEntity().level());
-            }
-            else
-                return;
+        if (shieldBlockFixedDamageAmount && event.getEntity().getUseItem().getItem() instanceof SPShieldItem) {
+            double baseBlockedDamage = ((SPShieldItem) event.getEntity().getUseItem().getItem()).getBlockedDamage(event.getEntity().getUseItem(), event.getEntity(), event.getEntity().level());
             float blockedDamage = (float) (baseBlockedDamage + ReinforcedEnchantment.getDamageBlocked(event.getEntity().getUseItem()));
             event.setBlockedDamage(blockedDamage);
         }
@@ -96,17 +72,6 @@ public class BaseFeature extends Feature {
 
         LightweightEnchantment.onTick(event.player);
         ShieldBashEnchantment.onTick(event.player);
-    }
-
-    @SubscribeEvent
-    public void onTooltip(ItemTooltipEvent event) {
-        if (!this.isEnabled()
-                || !shieldBlockFixedDamageAmount)
-            return;
-
-        if (event.getItemStack().is(Items.SHIELD)) {
-            SPShieldItem.addDamageBlockedText(event.getItemStack(), event.getToolTip(), SPShieldMaterials.IRON.damageBlocked);
-        }
     }
 
     public static boolean shouldRemoveShieldWindup() {