+++ src/main/java/insane96mcp/shieldsplus/setup/client/Client.java
@@ -3,6 +3,7 @@ package insane96mcp.shieldsplus.setup.client;
 import insane96mcp.shieldsplus.ShieldsPlus;
 import insane96mcp.shieldsplus.setup.SPItems;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.client.renderer.item.ItemPropertyFunction;
 import net.minecraft.world.item.*;
@@ -12,12 +13,18 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.registries.RegistryObject;
 
+import java.text.DecimalFormat;
+import java.text.DecimalFormatSymbols;
+
 @Mod.EventBusSubscriber(value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD, modid = ShieldsPlus.MOD_ID)
 public class Client {
 
     @SuppressWarnings("unused")
     public static void setup(final FMLClientSetupEvent event) {
         initShields();
+
+        DecimalFormatSymbols DECIMAL_FORMAT_SYMBOLS = new DecimalFormatSymbols(Minecraft.getInstance().getLocale());
+        ShieldsPlus.ONE_DECIMAL_FORMATTER = new DecimalFormat("#.#", DECIMAL_FORMAT_SYMBOLS);
     }
 
 /*    @SubscribeEvent