+++ src/main/java/insane96mcp/shieldsplus/world/item/SPShieldMaterial.java
@@ -8,25 +8,22 @@ import java.util.function.Supplier;
 
 public class SPShieldMaterial {
     public String materialName;
-    public double damageBlocked;
     public int durability;
     public Supplier<Item> repairItem;
     public TagKey<Item> repairTag;
     public int enchantmentValue;
     public Rarity rarity;
 
-    public SPShieldMaterial(String materialName, double damageBlocked, int durability, Supplier<Item> repairItem, int enchantmentValue, Rarity rarity) {
+    public SPShieldMaterial(String materialName, int durability, Supplier<Item> repairItem, int enchantmentValue, Rarity rarity) {
         this.materialName = materialName;
-        this.damageBlocked = damageBlocked;
         this.durability = durability;
         this.repairItem = repairItem;
         this.enchantmentValue = enchantmentValue;
         this.rarity = rarity;
     }
 
-    public SPShieldMaterial(String materialName, double damageBlocked, int durability, TagKey<Item> repairTag, int enchantmentValue, Rarity rarity) {
+    public SPShieldMaterial(String materialName, int durability, TagKey<Item> repairTag, int enchantmentValue, Rarity rarity) {
         this.materialName = materialName;
-        this.damageBlocked = damageBlocked;
         this.durability = durability;
         this.repairTag = repairTag;
         this.enchantmentValue = enchantmentValue;